Thu, May 4, 2023, 8:08 PM -04:00

+++ Week 4: Memory - PSET & Lab/filter-less/helpers.c
@@ -5,7 +5,7 @@
 void grayscale(int height, int width, RGBTRIPLE image[height][width])
 {
 
-   BYTE grayscaleHex;
+   int grayscaleHex;
 
    for (int i = 0; i < height; i++)
    {
